Discord Bot/XyPy/requirements.txt
discord
python-dotenv
Flask
matplotlib
sympy
wikipedia
forex-python